Now preview shows live webcam
The preview window will now display the live webcam. There are now two concepts of webcams, we have the scanning webcam and the live webcam.

The scanning webcam is the one that will constantly detect if it is being looked at, and the live webcam is the one that will eventually pipe through to a virtual webcam for use in video conferencing apps.