Fixing startup bug
When starting up, each cameras list of detections is set to False. On the first pass of the main program loop, this means that there is a high possibility that no best frame will be selected, causing an NRE when starting to prepare this non-existent frame for the virtual camera.

Now skipping cases where a best frame is not selected.