Adding webcam weightings
Now you can add a webcam to favour, and set the multiplication factor to the amount you want to weigh that camera over the other camera readings.

This is useful if you have a centre camera that you mainly look at, and it is more probable that you will be looking at that camera over other cameras.

It should be noted that the degree that the weighting has will also depend on the number of measurements taken. The more measurements, the more the weighting will multiply the readings. Also, if a camera has zero positive detections for a given measurement window, the weighting will have no effect as it will be multiplying against a zero.