Improving accuracy
Previously each measurement was a True/False, based on the method is_center. This method is based on a horizontal_ratio, that ranges from 0 (absolute left) to 1 (absolute right) where 0,.5 is looking dead center (at the camera). The is_center method checks for a horizontal ratio between 0.35 and 0.65, leaving lots of room for error.

Now we use the horizontal_ratio itself, and for each measurement we detect how far away the current horizontal_ratio is from dead center (0.5). This means that the lower the sum of the list of measurements, the higher the chance that a user is looking at the current camera.