Adding weighting for centre webcam
Slightly favouring cameras that are not the centre webcam, as this camera is expected to have a clearer view of someone at any given time so the results may be skewed towards it.

Still playing around with some ideas to try and make this more accurate. I'm going to see if it might be worth adding a calibration stage so that we get the average readings for each camera when a person is looking directly at them, so that we can make comparisons between the cameras more fair.